+++ frontend/lib/screens/home/reports_screen.dart
@@ -1280,7 +1280,7 @@ class _ReportsScreenState extends State<ReportsScreen> {
                       if (_damagedProductsReport != null) ...[
                         _metricCard(
                           'Damaged Items', 
-                          double.tryParse(_damagedProductsReport!['summary']['total_quantity_damaged']?.toString() ?? '0') ?? 0.0, 
+                          int.tryParse(_damagedProductsReport!['summary']['total_quantity_damaged']?.toString() ?? '0') ?? 0, 
                           color: Colors.orange[50], 
                           icon: Icons.warning,
                           isSmallMobile: isSmallMobile,
